# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Nami Transaction Tracking System.",
        "title": "Nami Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/actions": {
            "post": {
                "description": "Perform a domain-specific action",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actions"
                ],
                "summary": "Perform an action",
                "parameters": [
                    {
                        "description": "Action request",
                        "name": "action",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ActionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ActionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/fx/history": {
            "get": {
                "description": "Get daily FX rates for a period",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fx"
                ],
                "summary": "Get FX history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Base currency (default USD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Quote currency (default VND)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/fx/today": {
            "get": {
                "description": "Fetch today's FX rate for a currency pair, cache it, and return",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fx"
                ],
                "summary": "Get today's FX rate (fetched and stored)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Base currency (default USD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Quote currency (default VND)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/investments": {
            "get": {
                "description": "Retrieve investments based on filter criteria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investments"
                ],
                "summary": "Get investments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by asset",
                        "name": "asset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by account",
                        "name": "account",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by horizon",
                        "name": "horizon",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by open status",
                        "name": "is_open",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset results",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Investment"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/investments/available": {
            "get": {
                "description": "Retrieve open investments available for stake operations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investments"
                ],
                "summary": "Get available investments for stake",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset",
                        "name": "asset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account",
                        "name": "account",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Horizon",
                        "name": "horizon",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Investment"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/investments/stake": {
            "post": {
                "description": "Process a stake transaction and create/update investment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investments"
                ],
                "summary": "Create stake investment",
                "parameters": [
                    {
                        "description": "Stake transaction",
                        "name": "stake",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Investment"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/investments/summary": {
            "get": {
                "description": "Retrieve investment summary statistics",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investments"
                ],
                "summary": "Get investment summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by asset",
                        "name": "asset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by account",
                        "name": "account",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by horizon",
                        "name": "horizon",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by open status",
                        "name": "is_open",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.InvestmentSummary"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/investments/unstake": {
            "post": {
                "description": "Process an unstake transaction and update investment. The field `investment_id` is required and must reference an existing open investment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investments"
                ],
                "summary": "Create unstake investment",
                "parameters": [
                    {
                        "description": "Unstake transaction (requires investment_id)",
                        "name": "unstake",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Investment"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/prices/daily": {
            "get": {
                "description": "Get daily prices for a symbol or asset ID in a currency",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prices"
                ],
                "summary": "Get daily asset prices",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset symbol (e.g., BTC)",
                        "name": "symbol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Asset ID",
                        "name": "asset_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Target currency (default USD)",
                        "name": "currency",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reports/borrows/outstanding": {
            "get": {
                "description": "Get list of outstanding borrows as of a date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Get outstanding borrows",
                "parameters": [
                    {
                        "type": "string",
                        "description": "As of date (YYYY-MM-DD)",
                        "name": "as_of",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.OutstandingBorrow"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reports/cashflow": {
            "get": {
                "description": "Get cash flow over a period",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Get cash flow report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Days back from end_date",
                        "name": "days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CashFlowReport"
                        }
                    },
                    "400": {
                        "description": "Invalid period parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ActionRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "params": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "models.ActionResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "executed_at": {
                    "type": "string"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Transaction"
                    }
                }
            }
        },
...
        "models.InvestmentSummary": {
            "type": "object",
            "properties": {
                "closed_investments": {
                    "type": "integer"
                },
                "open_investments": {
                    "type": "integer"
                },
                "open_market_value": {
                    "type": "number"
                },
                "realized_pnl": {
                    "type": "number"
                },
                "roi_percent": {
                    "type": "number"
                },
                "total_deposits": {
                    "type": "number"
                },
                "total_investments": {
                    "type": "integer"
                },
                "total_pnl": {
                    "type": "number"
                },
                "total_withdrawals": {
                    "type": "number"
                },
                "unrealized_pnl": {
                    "type": "number"
                }
            }
        },
...
        "models.OutstandingBorrow": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "asset": {
                    "type": "string"
                }
            }
        },
...
    }
}